Can now swim properly
Use space to go up, ctrl to go down, and uhh WASD to swim around
there are 2 types of swimming, suface swimming and underwater swimming. on the survace you can only swim in the surface, if you then lower yourself using ctrl, you can swim in 4 directions, left, right and back are all based on the character. forward is based on the camera.